+++ CommonplaceBookApp/SavingTextEditViewController.swift
@@ -21,24 +21,6 @@ final class SavingTextEditViewController: UIViewController, TextEditViewControll
     super.init(nibName: nil, bundle: nil)
   }
 
-  override func viewDidLoad() {
-    super.viewDidLoad()
-    view.addSubview(textEditViewController.view)
-    textEditViewController.view.snp.makeConstraints { make in
-      make.edges.equalToSuperview()
-    }
-    addChild(textEditViewController)
-    textEditViewController.didMove(toParent: self)
-    textEditViewController.delegate = self
-    self.autosaveTimer = Timer.scheduledTimer(withTimeInterval: 1, repeats: true, block: { [weak self] _ in
-      if self?.hasUnsavedChanges ?? false { DDLogInfo("SavingTextEditViewController: autosave") }
-      self?.saveIfNeeded()
-    })
-    navigationItem.leftBarButtonItem = splitViewController?.displayModeButtonItem
-    navigationItem.leftItemsSupplementBackButton = true
-    navigationItem.largeTitleDisplayMode = .never
-  }
-
   required init?(coder: NSCoder) {
     fatalError("init(coder:) has not been implemented")
   }
@@ -57,6 +39,43 @@ final class SavingTextEditViewController: UIViewController, TextEditViewControll
       .map { return MarkdownAttributedStringRenderer.textOnly.render(node: $0) }
       .joined()
     navigationItem.title = attributedTitle.string
+  }
+
+  override func viewDidLoad() {
+    super.viewDidLoad()
+    view.addSubview(textEditViewController.view)
+    textEditViewController.view.snp.makeConstraints { make in
+      make.edges.equalToSuperview()
+    }
+    addChild(textEditViewController)
+    textEditViewController.didMove(toParent: self)
+    textEditViewController.delegate = self
+    self.autosaveTimer = Timer.scheduledTimer(withTimeInterval: 1, repeats: true, block: { [weak self] _ in
+      if self?.hasUnsavedChanges ?? false { DDLogInfo("SavingTextEditViewController: autosave") }
+      self?.saveIfNeeded()
+    })
+    navigationItem.largeTitleDisplayMode = .never
+  }
+
+  override func viewWillAppear(_ animated: Bool) {
+    super.viewWillAppear(animated)
+    navigationController?.hidesBarsOnSwipe = true
+    configureToolbar()
+  }
+
+  override func traitCollectionDidChange(_ previousTraitCollection: UITraitCollection?) {
+    configureToolbar()
+  }
+
+  private func configureToolbar() {
+    if splitViewController?.isCollapsed ?? false {
+      navigationController?.isToolbarHidden = false
+      toolbarItems = [UIBarButtonItem.flexibleSpace(), AppCommandsButtonItems.newNote()]
+    } else {
+      navigationItem.rightBarButtonItem = AppCommandsButtonItems.newNote()
+      navigationController?.isToolbarHidden = true
+      toolbarItems = []
+    }
   }
 
   /// Writes a note to storage.